move logic to mixin
